pyproject.toml
[project]
name = "meme-bot"
version = "0.1.0"
description = "A discord bot that generates memes based on conversation context"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py[voice]>=2.4.0",
    "logfire>=3.0.0",
    "pydantic>=2.10.4",
    "pydantic-ai>=0.0.18",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.6",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"